+++ src/main/java/app/freerouting/interactive/CopyItemState.java
@@ -178,11 +178,11 @@ public class CopyItemState extends InteractiveState {
       if (curr_cmp_no > 0) {
         // This item belongs to a component
         int new_cmp_no;
-        Integer curr_key = Integer.valueOf(curr_cmp_no);
+        Integer curr_key = curr_cmp_no;
         if (cmp_no_pairs.containsKey(curr_key)) {
           // the new component for this pin is already created
           Integer curr_value = cmp_no_pairs.get(curr_key);
-          new_cmp_no = curr_value.intValue();
+          new_cmp_no = curr_value;
         } else {
           Component old_component = board.components.get(curr_cmp_no);
           if (old_component == null) {
@@ -222,7 +222,7 @@ public class CopyItemState extends InteractiveState {
                   new_package);
           copied_components.add(new_component);
           new_cmp_no = new_component.no;
-          cmp_no_pairs.put(Integer.valueOf(curr_cmp_no), Integer.valueOf(new_cmp_no));
+          cmp_no_pairs.put(curr_cmp_no, new_cmp_no);
         }
         curr_item.assign_component_no(new_cmp_no);
       }
