+++ src/main/java/app/freerouting/interactive/DragItemState.java
@@ -126,11 +126,11 @@ public class DragItemState extends DragState {
         while (it.hasNext()) {
           Item curr_moved_item = it.next();
           for (int i = 0; i < curr_moved_item.net_count(); ++i) {
-            changed_nets.add(Integer.valueOf(curr_moved_item.get_net_no(i)));
+            changed_nets.add(curr_moved_item.get_net_no(i));
           }
         }
         for (Integer curr_net_no : changed_nets) {
-          hdlg.update_ratsnest(curr_net_no.intValue());
+          hdlg.update_ratsnest(curr_net_no);
         }
       }
     } else {
